Move log stats from handler package to logs package
