+++ logger/error_logger.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"log/slog"
 	"sync"
-
-	"github.com/chtc/chtc-go-logger/logger/handlers"
 )
 
 // Wrapper for ContextAwareErrorLogger that returns the latest log info with each logger call
@@ -14,7 +12,7 @@ type ContextAwareErrorLogger struct {
 	ContextAwareLogger
 }
 
-func (l *ContextAwareErrorLogger) Log(ctx context.Context, level slog.Level, msg string, attrs ...slog.Attr) handlers.LogStats {
+func (l *ContextAwareErrorLogger) Log(ctx context.Context, level slog.Level, msg string, attrs ...slog.Attr) LogStats {
 	l.mu.Lock()
 	defer l.mu.Unlock()
 
@@ -23,18 +21,18 @@ func (l *ContextAwareErrorLogger) Log(ctx context.Context, level slog.Level, msg
 }
 
 // Convenience methods for log levels
-func (l *ContextAwareErrorLogger) Info(ctx context.Context, msg string, attrs ...slog.Attr) handlers.LogStats {
+func (l *ContextAwareErrorLogger) Info(ctx context.Context, msg string, attrs ...slog.Attr) LogStats {
 	return l.Log(ctx, slog.LevelInfo, msg, attrs...)
 }
 
-func (l *ContextAwareErrorLogger) Debug(ctx context.Context, msg string, attrs ...slog.Attr) handlers.LogStats {
+func (l *ContextAwareErrorLogger) Debug(ctx context.Context, msg string, attrs ...slog.Attr) LogStats {
 	return l.Log(ctx, slog.LevelDebug, msg, attrs...)
 }
 
-func (l *ContextAwareErrorLogger) Warn(ctx context.Context, msg string, attrs ...slog.Attr) handlers.LogStats {
+func (l *ContextAwareErrorLogger) Warn(ctx context.Context, msg string, attrs ...slog.Attr) LogStats {
 	return l.Log(ctx, slog.LevelWarn, msg, attrs...)
 }
 
-func (l *ContextAwareErrorLogger) Error(ctx context.Context, msg string, attrs ...slog.Attr) handlers.LogStats {
+func (l *ContextAwareErrorLogger) Error(ctx context.Context, msg string, attrs ...slog.Attr) LogStats {
 	return l.Log(ctx, slog.LevelError, msg, attrs...)
 }
